refactor(Repository) : Repository 계층의 interface 도입
